fix(#689): Exponential data growth for large cardinality reflected references
If the project uses reflected references, this leads to exponential growth of the data files and slow indexing. The problem is that whenever the referencing entity is stored, the copy of the ReferenceStoragePart entity with the reflected reference is also stored (adding a single reflected reference to it).

+++ evita_store/evita_store_server/src/main/java/io/evitadb/store/catalog/DefaultCatalogPersistenceService.java
@@ -49,7 +49,7 @@ import io.evitadb.core.Catalog;
 import io.evitadb.core.CatalogVersionBeyondTheHorizonListener;
 import io.evitadb.core.EntityCollection;
 import io.evitadb.core.async.Scheduler;
-import io.evitadb.core.buffer.DataStoreIndexChanges;
+import io.evitadb.core.buffer.DataStoreChanges;
 import io.evitadb.core.buffer.DataStoreMemoryBuffer;
 import io.evitadb.core.buffer.WarmUpDataStoreMemoryBuffer;
 import io.evitadb.core.file.ExportFileService;
@@ -266,6 +266,13 @@ public class DefaultCatalogPersistenceService implements CatalogPersistenceServi
 	 * storage part persistence service for a given catalog version.
 	 */
 	private long[] catalogPersistenceServiceVersions;
+	/**
+	 * Contains information about cardinality of the warm-up version of the catalog - i.e. zero. This version is special
+	 * in the sense, it may be used repeatedly (version doesn't increment with catalog flushes) and because the array
+	 * {@link #catalogPersistenceServiceVersions} cannot contain multiple zeros, this counter is used to keep track of
+	 * the number of times the zero version was used.
+	 */
+	private int warmUpVersionCardinality;
 	/**
 	 * Contains the instance of {@link CatalogBootstrap} that contains the last bootstrap record that is currently used.
 	 */
@@ -862,6 +869,7 @@ public class DefaultCatalogPersistenceService implements CatalogPersistenceServi
 			)
 		);
 		this.catalogPersistenceServiceVersions = new long[]{catalogVersion};
+		this.warmUpVersionCardinality = 1;
 
 		if (lastCatalogBootstrap.fileLocation() == null) {
 			this.bootstrapUsed = recordBootstrap(catalogVersion, this.catalogName, 0, null);
@@ -945,6 +953,7 @@ public class DefaultCatalogPersistenceService implements CatalogPersistenceServi
 			catalogStoragePartPersistenceService
 		);
 		this.catalogPersistenceServiceVersions = new long[]{catalogVersion};
+		this.warmUpVersionCardinality = catalogVersion == 0 ? 1 : 0;
 
 		final File restoreFlagFile = catalogStoragePath.resolve(CatalogPersistenceService.RESTORE_FLAG).toFile();
 		verifyCatalogNameMatches(
@@ -1015,6 +1024,7 @@ public class DefaultCatalogPersistenceService implements CatalogPersistenceServi
 			catalogStoragePartPersistenceService
 		);
 		this.catalogPersistenceServiceVersions = new long[]{catalogVersion};
+		this.warmUpVersionCardinality = catalogVersion == 0 ? 1 : 0;
 
 		final CatalogHeader catalogHeader = catalogStoragePartPersistenceService.getCatalogHeader(catalogVersion);
 		this.entityCollectionPersistenceServices = CollectionUtils.createConcurrentHashMap(
@@ -1300,7 +1310,8 @@ public class DefaultCatalogPersistenceService implements CatalogPersistenceServi
 		long catalogVersion,
 		@Nonnull HeaderInfoSupplier headerInfoSupplier,
 		@Nonnull EntityCollectionHeader entityCollectionHeader,
-		@Nonnull DataStoreMemoryBuffer dataStoreBuffer) {
+		@Nonnull DataStoreMemoryBuffer dataStoreBuffer
+	) {
 		final CollectionFileReference collectionFileReference =
 			new CollectionFileReference(
 				entityCollectionHeader.entityType(),
@@ -1912,9 +1923,9 @@ public class DefaultCatalogPersistenceService implements CatalogPersistenceServi
 	}
 
 	@Override
-	public void flushTrappedUpdates(long catalogVersion, @Nonnull DataStoreIndexChanges dataStoreIndexChanges) {
+	public void flushTrappedUpdates(long catalogVersion, @Nonnull DataStoreChanges dataStoreChanges) {
 		// now store all the entity trapped updates
-		dataStoreIndexChanges.popTrappedUpdates()
+		dataStoreChanges.popTrappedUpdates()
 			.forEach(it -> getStoragePartPersistenceService(catalogVersion).putStoragePart(catalogVersion, it));
 	}
 
@@ -2058,11 +2069,23 @@ public class DefaultCatalogPersistenceService implements CatalogPersistenceServi
 					nonFlushedBlock -> this.reportNonFlushedContents(catalogName, nonFlushedBlock),
 					oldestRecordTimestamp -> DefaultCatalogPersistenceService.reportOldestHistoricalRecord(catalogName, oldestRecordTimestamp.orElse(null))
 				);
-				this.catalogStoragePartPersistenceService.put(
+				final CatalogOffsetIndexStoragePartPersistenceService previousService = this.catalogStoragePartPersistenceService.put(
 					catalogVersion,
 					newPersistenceService
 				);
-				this.catalogPersistenceServiceVersions = ArrayUtils.insertLongIntoOrderedArray(catalogVersion, this.catalogPersistenceServiceVersions);
+				if (previousService != null) {
+					previousService.close();
+					if (catalogVersion == 0) {
+						this.warmUpVersionCardinality++;
+					} else {
+						throw new GenericEvitaInternalError(
+							"Persistence storage instance is unexpectedly already registered!",
+							"Persistence storage instance for version `" + catalogVersion + "` is unexpectedly already registered!"
+						);
+					}
+				} else {
+					this.catalogPersistenceServiceVersions = ArrayUtils.insertLongIntoOrderedArray(catalogVersion, this.catalogPersistenceServiceVersions);
+				}
 
 				this.obsoleteFileMaintainer.removeFileWhenNotUsed(
 					catalogVersion,
@@ -2445,11 +2468,15 @@ public class DefaultCatalogPersistenceService implements CatalogPersistenceServi
 				lookupIndex >= 0 && lookupIndex < this.catalogPersistenceServiceVersions.length,
 				() -> new GenericEvitaInternalError("Catalog version " + catalogVersion + " not found in the catalog persistence service versions!")
 			);
-			final long versionToRemove = this.catalogPersistenceServiceVersions[lookupIndex];
-			this.catalogPersistenceServiceVersions = ArrayUtils.removeLongFromArrayOnIndex(this.catalogPersistenceServiceVersions, lookupIndex);
-			// remove the service and release its resources
-			final CatalogOffsetIndexStoragePartPersistenceService storageService = this.catalogStoragePartPersistenceService.remove(versionToRemove);
-			storageService.close();
+			if (catalogVersion == 0 && this.warmUpVersionCardinality > 0) {
+				this.warmUpVersionCardinality--;
+			} else {
+				final long versionToRemove = this.catalogPersistenceServiceVersions[lookupIndex];
+				this.catalogPersistenceServiceVersions = ArrayUtils.removeLongFromArrayOnIndex(this.catalogPersistenceServiceVersions, lookupIndex);
+				// remove the service and release its resources
+				final CatalogOffsetIndexStoragePartPersistenceService storageService = this.catalogStoragePartPersistenceService.remove(versionToRemove);
+				storageService.close();
+			}
 		} catch (InterruptedException e) {
 			throw new GenericEvitaInternalError(
 				"Failed to lock the catalog persistence service for catalog `" + this.catalogName + "`!",
